requirements.txt
Django==1.10
djangorestframework==3.4.7
pycountry==16.11.27.1
